+++ src/constants/property-metadata.constants.ts
@@ -1177,6 +1177,14 @@ export const propertyMetadata = {
     description: "Full description for the source category type.",
     example: "Cogeneration",
   },
+  stackPipeId: {
+    fieldLabels: {
+      label: "",
+      value: "stackPipeId"
+    },
+    description: "This is the alphanumeric code assigned by a source to identify a multiple or common stack or pipe at which emissions are determined.",
+    example: "CS0AAN"
+  },
   startBlock: {
     fieldLabels: {
       label: "Serial Number Start",
@@ -1234,6 +1242,22 @@ export const propertyMetadata = {
     },
     description: "Sum of operating time for the time interval",
     example: 1795.25,
+  },
+  testSummaryId: {
+    fieldLabels: {
+      label: "",
+      value: "id"
+    },
+    description: "Unique identifier of a test summary record.",
+    example: "TESTSUM7DAY123"
+  },
+  testTypeCode: {
+    fieldLabels: {
+      label: "",
+      value: "testTypeCode",
+    },
+    description: "Code used to identify test type or the type of test performed.",
+    example: "7DAY"
   },
   totalAllowancesDeducted: {
     fieldLabels: {
@@ -2640,7 +2664,7 @@ export const propertyMetadata = {
     description: "Flag for active monitor location record",
     example: true,
   },
-  
+
   monitorPlanCommentDTOId:{
     fieldLabels: {
       label: "",
@@ -2693,6 +2717,60 @@ export const propertyMetadata = {
     },
     description: "Comment on a monitoring plan",
     example: "DAHS was updated for GT-1 on 11/10/2016 by a previous owner, but the component ID was not updated on the gas system. This was corrected on 9/8/2020.",
+  },
+  
+  monitorPlanReportingFreqDTOId:{
+    fieldLabels: {
+      label: "",
+      value: "id",
+    },
+    description: "Unique identifier of a monitoring plan Reporting Frequency record.",
+    example: "MDC-7A8D8E101C484732954BE84D49D9249A",
+  },
+  
+  monitorPlanReportingFreqDTOPlanId:{
+    fieldLabels: {
+      label: "",
+      value: "planId",
+    },
+    description: "Unique identifier of a monitoring plan record.",
+    example: "MDC-2483BA898490441B97B7D7CFB1BE6D9A",
+  },
+  
+  monitorPlanReportingFreqDTOUserId:{
+    fieldLabels: {
+      label: "",
+      value: "userId",
+    },
+    description: "User account or source of data that added or updated record.",
+    example: "abcde",
+  },
+  
+  monitorPlanReportingFreqDTOAddDate:{
+    fieldLabels: {
+      label: "",
+      value: "addDate",
+    },
+    description: "Date and time in which record was added.",
+    example: "2020-10-28",
+  },
+  
+  monitorPlanReportingFreqDTOUpdateDate:{
+    fieldLabels: {
+      label: "",
+      value: "updateDate",
+    },
+    description: "Date and time in which record was last updated.",
+    example: "2020-10-28",
+  },
+  
+  monitorPlanReportingFreqDTOMonitoringPlanReportingFreqCode:{
+    fieldLabels: {
+      label: "",
+      value: "reportFrequencyCode",
+    },
+    description: "Report Frequency Code on a monitoring plan",
+    example: "OS",
   },
   
   monitorPlanCommentDTOBeginDate:{
@@ -4315,6 +4393,15 @@ unitCapacityDTOAddDate:{
     example: "1,2,CSA00N",
   },
   
+  monitorPlanDTOBeginReportPeriodId:{
+    fieldLabels: {
+      label: "",
+      value: "beginReportPeriodId",
+    },
+    description: "Unique identifier of a reporting period record.",
+    example: null,
+  },
+
   monitorPlanDTOEndReportPeriodId:{
     fieldLabels: {
       label: "",